cheese_chatbot/requirements.txt
langchain
langchain-openai
langgraph
streamlit
pymongo
pinecone-client
fastapi
uvicorn
requests
python-dotenv
tiktoken 